Generalize version constraint in simple test
Prior to this change, the simple terratest was pinned to a specific
version of the aws-logs module. Generalizing the version constraint will
allow this module to more easily test minor and patch version updates to
the aws-logs module.

This change modifies the terraform module version constraint in the
simple terratest to only pin to the major version, rather than pin to a
specific major/minor/patch version.

https://www.pivotaltracker.com/story/show/169409905