Cleanup simple terratest
Prior to this change, referencing module.logs.aws_logs_bucket directly
in the cloudtrail module terratest would fail with an
InsufficientS3BucketPolicyException due to a race condition around how
the aws-logs module attached the bucket policy. v4.1.0 of the aws-logs
module fixed this issue, so the workaround in the cloudtrail terratest
is no longer needed.

This change removes an earlier workaround and references
module.logs.aws_logs_bucket directly in the cloudtrail module test.

https://www.pivotaltracker.com/story/show/169792025